discord dot com more like PEE

+++ index.js
@@ -287,7 +287,7 @@ client.on("messageUpdate", async (oldMessage, newMessage) => {
 				console.log("Something went wrong when fetching the message: ", error);
 			});
 	}
-	if (inviteSites.some(site => newMessage.content.includes(site)) && config.moderation.automod.allowInviteLinks == false) {
+	if (newMessage && newMessage.content && inviteSites.some(site => newMessage.content.includes(site)) && config.moderation.automod.allowInviteLinks == false) {
 		newMessage.delete();
 		newMessage.channel.send(format(vukkytils.getString("DISCORD_INVITES_DISABLED_AUTOMOD"), newMessage.author)).then(msg => setTimeout(() => msg.delete(), 7000));
 	}
